% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxscore.R
\name{boxscore}
\alias{boxscore}
\alias{get_boxscore}
\alias{parse_boxscore}
\alias{build_season_boxscores_df}
\title{functions for working with boxscore data}
\usage{
get_boxscore(game_id)

parse_boxscore(boxscore_json)

build_season_boxscores_df(season, sleep_time = 5, verbose = FALSE)
}
\arguments{
\item{game_id}{character string of game id}

\item{boxscore_json}{list output from \code{get_boxscore()}}

\item{season}{character string of season, which can be fussy - for instance, the 2023-24 season shows up as both \code{202324} and \code{20232024} in some use cases}

\item{sleep_time}{numeric how long to sleep between API calls in seconds}

\item{verbose}{logical should functions with the option to print messages print them}
}
\description{
functions for working with boxscore data
}
\details{
\code{get_boxscore()} returns a list object from the request response

\code{parse_boxscore()} takes a list from \code{get_boxscore()} as input, and returns a data.frame of boxscores

\code{build_season_boxscores_df()} is a wrapper that takes a season string as input and builds a data.frame with all the boxscores for that season from scratch. THIS IS V TIME CONSUMING AND COULD MAKE THE API GODS MAD, USE A GOOD SLEEP TIME HERE.
}
