% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.r
\name{get_game_feed}
\alias{get_game_feed}
\alias{get_game_content}
\title{NHL Games}
\usage{
get_game_feed(game_id)

get_game_content(game_id)
}
\arguments{
\item{game_id}{game id to request}
}
\description{
Query data on NHL games, including stats and content.
}
\examples{
\dontrun{
game <- get_game_feed("2017020449")
}
}
