% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.r
\name{get_schedule}
\alias{get_schedule}
\title{NHL Schedule}
\usage{
get_schedule(start_date = Sys.Date(), end_date = Sys.Date())
}
\arguments{
\item{start_date, end_date}{the start and end dates to get schedule for}
}
\description{
Query data on NHL schedule.
}
\examples{
\dontrun{
schedule <- get_schedule("2017-12-12", "2017-12-15")
}
}
