% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.r
\name{get_teams}
\alias{get_teams}
\alias{get_team}
\alias{get_team_stats}
\alias{get_roster}
\title{NHL Teams}
\usage{
get_teams()

get_team(team_id)

get_team_stats(team_id)

get_roster(team_id)
}
\arguments{
\item{team_id}{team id to request}
}
\description{
Query data on NHL teams, including rosters.
}
\examples{
\dontrun{
teams <- get_teams()
team1 <- get_team("1")
roster1 <- get_roster("1")
}
}
