% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_awards.R
\name{nhl_awards}
\alias{nhl_awards}
\title{Retrieve metadata on NHL awards from the API}
\usage{
nhl_awards(awardIds = NULL)
}
\arguments{
\item{awardIds}{\code{integer()}, vector of one or more award ids
or \code{NULL} (default) for all awards. The current set of valid
ids seems to be \code{1:24}.}
}
\value{
\code{data.frame}, with information on awards, one row per award.
}
\description{
Retrieve metadata on NHL awards from the API
}
\examples{
\dontrun{
  # Get information on current draft
  nhl_awards()

  # Get information on 3 historical awards
  nhl_awards(1:3)
}
}
