% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_draft_prospects.R
\name{nhl_draft_prospects}
\alias{nhl_draft_prospects}
\title{Retrieve metadata on NHL draft prospects from the API}
\usage{
nhl_draft_prospects(prospectIds = NULL)
}
\arguments{
\item{prospectIds}{\code{integer()}, vector of one or more ids of draft
prospects or \code{NULL} (default) for all exposed prospects.}
}
\value{
\code{data.frame}, with information on draft prospects, one row
per draft prospect.
}
\description{
Retrieve metadata on NHL draft prospects from the API
}
\examples{
\dontrun{
  # Get information on current draft prospects
  nhl_draft_prospects()
}
}
