% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_drafts.R
\name{nhl_drafts}
\alias{nhl_drafts}
\title{Retrieve metadata on NHL drafts from the API}
\usage{
nhl_drafts(draftYears = NULL)
}
\arguments{
\item{draftYears}{\code{integer()}, vector of one or more years in \code{YYYY}
format or \code{NULL} (default) for the current year's draft. Also
accepts a character vector of years in \code{YYYY} format.}
}
\value{
\code{data.frame}, with information on drafts, one row
year.
}
\description{
Retrieve metadata on NHL drafts from the API
}
\examples{
\dontrun{
  # Get information on current draft
  nhl_drafts()

  # Get information on 3 historical drafts
  nhl_drafts(2015:2017)
}
}
