% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_get_data.R
\name{nhl_get_data}
\alias{nhl_get_data}
\title{Get data from the API}
\usage{
nhl_get_data(sUrl, silent = getOption("nhlapi_try_silent"))
}
\arguments{
\item{sUrl}{\code{character(1)}, the URL to get the data from.}

\item{silent}{\code{logical(1)}, passed to \code{try()}.}
}
\value{
\code{list} with the retrieved data or class \code{try-error}.
}
\description{
Gets data from the NHL API using \code{jsonlite}.
}
\seealso{
\code{\link[=nhl_url]{nhl_url()}}
}
