% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_get_data.R
\name{nhl_get_datas}
\alias{nhl_get_datas}
\title{Get data from the API for multiple urls}
\usage{
nhl_get_datas(urls)
}
\arguments{
\item{urls}{\code{character()}, vector of urls to retrieve
the data from.}
}
\value{
\code{list} of results retrieved using \code{\link[=nhl_get_data]{nhl_get_data()}}.
}
\description{
Get data from the API for multiple urls
}
