% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_md_endpoints.R
\name{nhl_md_event_types}
\alias{nhl_md_event_types}
\title{Get event types metadata}
\usage{
nhl_md_event_types()
}
\value{
\code{list}, with metadata on event types.
}
\description{
Get event types metadata
}
