% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_md_endpoints.R
\name{nhl_md_game_statuses}
\alias{nhl_md_game_statuses}
\title{Get game status metadata}
\usage{
nhl_md_game_statuses()
}
\value{
\code{list}, with metadata on game statuses.
}
\description{
Get game status metadata
}
