% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_md_endpoints.R
\name{nhl_md_game_types}
\alias{nhl_md_game_types}
\title{Get game type metadata}
\usage{
nhl_md_game_types()
}
\value{
\code{list}, with metadata on game types.
}
\description{
Get game type metadata
}
