% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_md_endpoints.R
\name{nhl_md_play_types}
\alias{nhl_md_play_types}
\title{Get play types metadata}
\usage{
nhl_md_play_types()
}
\value{
\code{list}, with metadata on play types.
}
\description{
Get play types metadata
}
