% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_md_endpoints.R
\name{nhl_md_standings_types}
\alias{nhl_md_standings_types}
\title{Get standings types metadata}
\usage{
nhl_md_standings_types()
}
\value{
\code{list}, with metadata on standings types.
}
\description{
Get standings types metadata
}
