% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_md_endpoints.R
\name{nhl_md_stat_types}
\alias{nhl_md_stat_types}
\title{Get stat types metadata}
\usage{
nhl_md_stat_types()
}
\value{
\code{list}, with metadata on stat types.
}
\description{
Get stat types metadata
}
