% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_md_endpoints.R
\name{nhl_md_tournament_types}
\alias{nhl_md_tournament_types}
\title{Get tournament types metadata}
\usage{
nhl_md_tournament_types()
}
\value{
\code{list}, with metadata on tournament types.
}
\description{
Get tournament types metadata
}
