% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_players_seasons.R
\name{nhl_players_allseasons}
\alias{nhl_players_allseasons}
\title{Retrieve all seasons statistics for players}
\usage{
nhl_players_allseasons(playerNames, playerIds = NULL)
}
\arguments{
\item{playerNames}{\code{character()}, vector of one or more player names.
Not case sensitive for convenience.}

\item{playerIds}{\code{integer()}, vector of one or more ids of the
players. The ids correspond to the ids expected by the NHL
API's people endpoint. For most cases the \code{playerNames} argument
can be provided for more convenient usage.}
}
\value{
data.frame
}
\description{
Retrieve all seasons statistics for players
}
\examples{
\dontrun{

 # With player names
 nhl_players_allseasons(c("joe sakic", "Peter Forsberg"))

 # With player ids
 nhl_players_allseasons(c(8451101, 8458554))
}
}
