% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_players_seasons.R
\name{nhl_players_seasons}
\alias{nhl_players_seasons}
\title{Retrieve selected seasons statistics for players}
\usage{
nhl_players_seasons(playerNames, seasons, playerIds = NULL,
  playoffs = FALSE)
}
\arguments{
\item{playerNames}{\code{character()}, player names, not case
sensitive for convenience.}

\item{seasons}{\code{character()}, seasons in the format \code{"YYYYZZZZ"},
where ZZZZ = YYYY + 1, e.g. \code{"19951996"}.}

\item{playerIds}{\code{integer()}, ids of the players.}

\item{playoffs}{\code{logical(1)} if \code{FALSE} (default) get the regular
seasons data, if \code{TRUE}, get the data for the playoffs.}
}
\value{
data.frame
}
\description{
Retrieve selected seasons statistics for players
}
\examples{
\dontrun{
 nhl_players_seasons(
   playerIds = c(8451101, 8458554),
   seasons = "19951996",
   playoffs = TRUE
 )
}
}
