% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_retrieve_players.R
\name{nhl_retrieve_players}
\alias{nhl_retrieve_players}
\title{Retrieve metadata for players}
\usage{
nhl_retrieve_players(playerIds)
}
\arguments{
\item{playerIds}{\code{integer()}, ids of the players.}
}
\value{
data.frame
}
\description{
Retrieve metadata for players
}
\examples{
\dontrun{
 nhl_retrieve_players(c(8451101, 8458554))
}
}
