% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_players_seasons.R
\name{nhl_retrieve_players_allseasons}
\alias{nhl_retrieve_players_allseasons}
\title{Retrieve all seasons statistics for players}
\usage{
nhl_retrieve_players_allseasons(playerIds)
}
\arguments{
\item{playerIds}{\code{integer()}, ids of the players.}
}
\value{
data.frame
}
\description{
Retrieve all seasons statistics for players
}
\examples{
\dontrun{
 nhl_retrieve_players_allseasons(c(8451101, 8458554))
}
}
