% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_retrieve_players_seasons.R
\name{nhl_retrieve_players_seasons}
\alias{nhl_retrieve_players_seasons}
\title{Retrieve selected seasons statistics for players}
\usage{
nhl_retrieve_players_seasons(playerIds, seasons, playoffs = FALSE)
}
\arguments{
\item{playerIds}{\code{integer()}, ids of the players.}

\item{seasons}{\code{character()}, seasons in the format \code{"YYYYZZZZ"},
where ZZZZ = YYYY + 1, e.g. \code{"19951996"}.}

\item{playoffs}{\code{logical(1)} if \code{FALSE} (default) get the regular
seasons data, if \code{TRUE}, get the data for the playoffs.}
}
\value{
data.frame
}
\description{
Retrieve selected seasons statistics for players
}
\examples{
\dontrun{
 nhl_retrieve_players_seasons(c(8451101, 8458554), "19951996", TRUE)
}
}
