% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_teams.R
\name{nhl_teams_rosters}
\alias{nhl_teams_rosters}
\title{Get rosters for teams}
\usage{
nhl_teams_rosters(teamIds = NULL, seasons = NULL)
}
\arguments{
\item{teamIds}{\code{integer()}, ids of the teams or \code{NULL} (default)
for all teams. As of end of 2019, the valid team ids seem to be
in the \code{1:54} range.}

\item{seasons}{\code{character()}, seasons in the format \code{"YYYYZZZZ"},
where ZZZZ = YYYY + 1, e.g. \code{"19951996"}.}
}
\value{
data.frame, with an element called \code{roster.roster}
that in itself is a data.frame with the roster data.
}
\description{
Get rosters for teams
}
\examples{
 \dontrun{
  # Current rosters for all teams
  nhl_teams_rosters()

  # Rosters for all teams for past seasons
  nhl_teams_rosters(seasons = c("19931994", "19931994"))

  # Roster for Devils and Islanders
  nhl_teams_rosters(
    teamIds = 1:2,
    seasons = c("19931994", "19931994")
  )
}
}
