% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_teams.R
\name{nhl_teams_shedule_previous}
\alias{nhl_teams_shedule_previous}
\title{Get details for the teams' previous game}
\usage{
nhl_teams_shedule_previous(teamIds = NULL)
}
\arguments{
\item{teamIds}{\code{integer()}, ids of the teams or \code{NULL} (default)
for all teams. As of end of 2019, the valid team ids seem to be
in the \code{1:54} range.}
}
\value{
data.frame, with elements with names starting with
\code{previousGameSchedule} that contain data on the teams' previous
game. One row per team.
}
\description{
Get details for the teams' previous game
}
\examples{
 \dontrun{
  # Next game for all teams
  nhl_teams_shedule_previous()

  # Next game for selected teams
  nhl_teams_shedule_previous(c(1,3,5))
}
}
