% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_url.R
\name{nhl_url_add_params}
\alias{nhl_url_add_params}
\title{Add parameters to urls}
\usage{
nhl_url_add_params(url, params)
}
\arguments{
\item{url}{\code{character()} vector of urls.}

\item{params}{\code{named list()} of parameters that will be
concatenated to the end of the urls after \code{?}. Parameters
can have multiple values, in which case multiple urls
are created.}
}
\value{
\code{character()} of the same length as all the
combinations of \code{url} and \code{params}.
}
\description{
Add parameters to urls
}
