% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_awards.R
\name{nhl_url_awards}
\alias{nhl_url_awards}
\title{Create an NHL API URL for awards}
\usage{
nhl_url_awards(awardIds = NULL)
}
\arguments{
\item{awardIds}{\code{integer()}, vector of one or more award ids
or \code{NULL} (default) for all awards. The current set of valid
ids seems to be \code{1:24}.}
}
\value{
\code{character()} of same length as \code{awardIds} or length \code{1}
if \code{awardIds} is \code{NULL}.
}
\description{
Create an NHL API URL for awards
}
\examples{
\dontrun{
  nhl_url_awards()
  nhl_url_awards(1:3)
}

}
