% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_conferences.R
\name{nhl_url_conferences}
\alias{nhl_url_conferences}
\title{Create an NHL API URL for conferences}
\usage{
nhl_url_conferences(conferenceIds = NULL)
}
\arguments{
\item{conferenceIds}{\code{integer()}, ids of the conferences or \code{NULL}
(default) for all conferences As of end of 2019, the valid
conference ids seem to be in the \code{1:7} range.}
}
\value{
\code{character()}, API URLs, same length as \code{teamIds} or
length \code{1} if \code{teamIds} is \code{NULL}.
}
\description{
Create an NHL API URL for conferences
}
\examples{
  nhlapi:::nhl_url_conferences()
  nhlapi:::nhl_url_conferences(1:3)

}
