% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_divisions.R
\name{nhl_url_divisions}
\alias{nhl_url_divisions}
\title{Create an NHL API URL for divisions}
\usage{
nhl_url_divisions(divisionIds = NULL)
}
\arguments{
\item{divisionIds}{\code{integer()}, ids of the teams or \code{NULL} (default)
for all teams. As of end of 2019, the valid division ids seem to be
in the \code{15:18} range.}
}
\value{
\code{character()} of same length as \code{teamIds} or length \code{1}
if \code{teamIds} is \code{NULL}
}
\description{
Create an NHL API URL for divisions
}
\examples{
\dontrun{
  nhl_url_divisions()
  nhl_url_divisions(1:3)
}

}
