% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_draft_prospects.R
\name{nhl_url_draft_prospects}
\alias{nhl_url_draft_prospects}
\title{Create an NHL API URL for draft prospects}
\usage{
nhl_url_draft_prospects(prospectIds = NULL)
}
\arguments{
\item{prospectIds}{\code{integer()}, vector of one or more ids of
draft prospects or \code{NULL} (default) for all exposed prospects.}
}
\value{
\code{character()}, API URLs, same length as \code{prospectIds} or
length \code{1} if \code{prospectIds} is \code{NULL}.
}
\description{
Create an NHL API URL for draft prospects
}
\examples{
  nhlapi:::nhl_url_draft_prospects()

}
