% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_players.R
\name{nhl_url_players}
\alias{nhl_url_players}
\title{Create an NHL API URL for players}
\usage{
nhl_url_players(playerIds)
}
\arguments{
\item{playerIds}{\code{integer()}, vector of one or more ids of the
players. The ids correspond to the ids expected by the NHL
API people endpoint. For most cases the \code{playerNames} argument
can be provided for more convenient usage.}
}
\value{
\code{character()} of same length as \code{playerIds}
}
\description{
Create an NHL API URL for players
}
\examples{
\dontrun{
  nhl_url_players(playerIds = c(8477474, 8477475))
}

}
