% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_url_players_seasons.R
\name{nhl_url_players_seasons}
\alias{nhl_url_players_seasons}
\title{Create an NHL API URL for  players' seasons statistics}
\usage{
nhl_url_players_seasons(playerIds, seasons, playoffs = FALSE)
}
\arguments{
\item{playerIds}{\code{integer()}, ids of the players.}

\item{seasons}{\code{character()}, seasons in the format \code{"YYYYZZZZ"},
where ZZZZ = YYYY + 1, e.g. \code{"19951996"}.}

\item{playoffs}{\code{logical(1)} if \code{FALSE} (default) get the regular
seasons data, if \code{TRUE}, get the data for the playoffs.}
}
\description{
Create an NHL API URL for  players' seasons statistics
}
\details{
If multiple players and seasons are provided, urls will
be created for all combinations of players and seasons.
}
\examples{
\dontrun{
  # Joe Sakic, regular season 1995/1996
  nhl_url_players_seasons(8451101L, "19951996")
  # Joe Sakic, playoffs 1995/1996 and 1996/1997
  nhl_url_players_seasons(
    8451101L,
    c("19951996", "19961997"),
    playoffs = TRUE
   )
}
}
