% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_url_players_stats.R
\name{nhl_url_players_stats}
\alias{nhl_url_players_stats}
\title{Create an NHL API stats URL for players}
\usage{
nhl_url_players_stats(playerIds)
}
\arguments{
\item{playerIds}{\code{integer()}, ids of the players.}
}
\value{
\code{character()} of same length as \code{playerIds}
}
\description{
Create an NHL API stats URL for players
}
\examples{
\dontrun{nhl_url_players_stats(8477474)}

}
