% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_seasons.R
\name{nhl_url_seasons}
\alias{nhl_url_seasons}
\title{Create an NHL API URL for seasons}
\usage{
nhl_url_seasons(seasons = NULL)
}
\arguments{
\item{seasons}{\code{numeric()}, \code{integer()} or \code{character()},
vector of starting years of desired seasons in \code{YYYY}
format, e.g. \code{1995} or \code{"1995"} for season 1995-1996.
Accepts vectors such as \code{c(1995:2000, 2010)} to generate
multiple seasons.

Alternatively, also accepts \code{character()} with seasons in the
format \code{"YYYYZZZZ"}, where \code{ZZZZ = YYYY + 1}, e.g. \code{"19951996"}.
This is the format that ultimately gets sent to the NHL API.

Some API endpoints, notably \code{seasons} exposed via \code{\link[=nhl_seasons]{nhl_seasons()}}
also allow the value \code{"current"} to passed. This value will be
returned unchanged.}
}
\value{
\code{character()}, of API URLs, same length as \code{seasons}
or length \code{1} if \code{seasons} is \code{NULL}.
}
\description{
Create an NHL API URL for seasons
}
\examples{
\dontrun{
  nhl_url_seasons()
  nhl_url_seasons(2015:2017)
  nhl_url_seasons("20152016")
}

}
