% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_venues.R
\name{nhl_url_venues}
\alias{nhl_url_venues}
\title{Create an NHL API URL for venues}
\usage{
nhl_url_venues(venueIds = NULL)
}
\arguments{
\item{venueIds}{\code{integer()}, vector of one or more venue ids
or \code{NULL} (default) for all currently exposed venues. The
exported values seem incomplete, so it may be worth it to
investigate other ids.}
}
\value{
\code{character()}, API URLs, same length as \code{venueIds} or
length \code{1} if \code{venueIds} is \code{NULL}.
}
\description{
Create an NHL API URL for venues
}
\examples{
\dontrun{
  nhl_url_venues()
  nhl_url_venues(5000:5006)
}

}
