% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_venues.R
\name{nhl_venues}
\alias{nhl_venues}
\title{Retrieve metadata on NHL venues from the API}
\usage{
nhl_venues(venueIds = NULL)
}
\arguments{
\item{venueIds}{\code{integer()}, vector of one or more venue ids
or \code{NULL} (default) for all currently exposed venues. The
exported values seem incomplete, so it may be worth it to
investigate other ids.}
}
\value{
\code{data.frame}, with information on venues, one row per venue.
}
\description{
Retrieve metadata on NHL venues from the API
}
\examples{
\dontrun{
  # Get information on currently exposed venues
  nhl_venues()

  # Get information on 3 historical venues
  nhl_venues(5000:5006)
}
}
