% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp.R
\name{pbp}
\alias{pbp}
\alias{get_pbp}
\alias{parse_pbp}
\title{functions for working with play by play data}
\usage{
get_pbp(game_id)

parse_pbp(pbp_json, verbose = FALSE)
}
\arguments{
\item{game_id}{character string of game id}

\item{pbp_json}{list output from \code{get_pbp()}}
}
\description{
functions for working with play by play data
}
\details{
\code{get_pbp()} returns a list object from the request response

\code{parse_pbp()} takes a list from \code{get_pbp()} as input, and returns a data.frame of plays
}
