% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_plot_utils.R
\name{plot_rink}
\alias{plot_rink}
\title{Plot an NHL rink}
\usage{
plot_rink()
}
\description{
Initialize a plot in base graphics with a to-scale NHL rink as the background
}
\details{
The placement of rink features & their sizes are exact according to the NHL rulebook; see citation.
}
\references{
\url{https://nhl.bamcontent.com/images/assets/binary/308893668/binary-file/file.pdf}
}
