% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_utils.R
\name{util_convert_minsonice}
\alias{util_convert_minsonice}
\title{Convert \code{"mm:ss"} character to numeric minutes}
\usage{
util_convert_minsonice(chr, splitter = ":")
}
\arguments{
\item{chr}{\code{character()}, vector in format \code{"mins:secs"}.}

\item{splitter}{\code{character(1)}, that splits
minutes and seconds in elements of \code{chr}.}
}
\value{
\code{numeric()}, vector of times in minutes. Same length
as \code{chr}.
}
\description{
Convert \code{"mm:ss"} character to numeric minutes
}
\examples{
   nhlapi:::util_convert_minsonice(c("20:00", "1500:30"))
}
