% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_utils.R
\name{util_flatten_list}
\alias{util_flatten_list}
\title{Flatten a list into a data frame keeping classes}
\usage{
util_flatten_list(x)
}
\arguments{
\item{x}{\code{list()}, note that a \code{data.frame} is also a \code{list}.}
}
\value{
data.frame with the list flattened and classes preserved.
}
\description{
DEPRECATED, NOT USED, TO BE REMOVED.
}
\examples{
\dontrun{
  x <- list(a = 1L, b = list(1, list(2L, TRUE)))
  nhl_util_flatten_list(x)
 }

}
