% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_utils.R
\name{util_generate_sysdata}
\alias{util_generate_sysdata}
\title{Generate the sysdata.rda file}
\usage{
util_generate_sysdata(playerIds = 8444849L:8490000L,
  tgtPath = "sysdata.rda")
}
\arguments{
\item{playerIds}{\code{integer()}, vector of playerIds.}

\item{tgtPath}{\code{character(1)}, path where to save
the generated object, \code{NULL} to not save.}
}
\value{
\code{data.frame} with player name hashes and ids.
}
\description{
Generate the sysdata.rda file
}
