% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_utils.R
\name{util_inherit_attributes}
\alias{util_inherit_attributes}
\title{Inherit attributes from another object}
\usage{
util_inherit_attributes(src, tgt, atrs = c("url", "copyright"))
}
\arguments{
\item{src}{\code{object}, with attributes to be inherited by \code{tgt}.}

\item{tgt}{\code{object}, onto which attributes of \code{src} should be added.}

\item{atrs}{\code{character()}, vector of names of attributes
of \code{src} to be added to \code{tgt}.}
}
\value{
\code{object} same as \code{tgt} with attributes added
}
\description{
Take attributes with names specified by \code{atrs}
from object \code{src} and adds them as the same attributes to \code{tgt}.
}
