% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_utils.R
\name{util_process_copyright}
\alias{util_process_copyright}
\title{Move copyright information to attribute}
\usage{
util_process_copyright(x, el = "copyright")
}
\arguments{
\item{x}{\code{list()}, to be processed.}

\item{el}{\code{character(1)}, name of the element to remove.
Defaults to \code{"copyright"} as this is the intended use
of the function.}
}
\value{
\code{list} with the \code{el} element removed and added
as attribute, if it is present in \code{x}. Unchanged \code{x}
otherwise.
}
\description{
Removes the element named \code{el} from \code{x} if
present and keeps the information as an equally named
attribute.
}
