% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_conferences.R
\name{nhl_conferences}
\alias{nhl_conferences}
\title{Retrieve metadata on NHL conferences from the API}
\usage{
nhl_conferences(conferenceIds = NULL)
}
\arguments{
\item{conferenceIds}{\code{integer()}, ids of the conferences or \code{NULL}
(default) for all conferences As of end of 2019, the valid
conference ids seem to be in the \code{1:7} range.}
}
\value{
\code{data.frame}, with information on conferences, one row
per conference.
}
\description{
Retrieve metadata on NHL conferences from the API
}
\examples{
\dontrun{
  # Get information on all conferences
  nhl_conferences()

  # Get information on 2 selected conferences
  nhl_conferences(5:6)
}
}
