% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_get_data.R
\name{nhl_get_data}
\alias{nhl_get_data}
\title{Get data from the API for multiple urls}
\usage{
nhl_get_data(urls, flatten = TRUE)
}
\arguments{
\item{urls}{\code{character()}, vector of urls to retrieve
the data from.}

\item{flatten}{\code{logical(1)}, if \code{TRUE} (default) automatically
flattens nested data frames into a single non-nested data frame.}
}
\value{
\code{list} of results retrieved using \code{\link[=nhl_get_data_worker]{nhl_get_data_worker()}}.
}
\description{
Get data from the API for multiple urls
}
