% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_utils.R
\name{nhl_make_seasons}
\alias{nhl_make_seasons}
\title{Make a vector of seasons consumable by the API}
\usage{
nhl_make_seasons(years = 1950L:2019L)
}
\arguments{
\item{years}{\code{integer()}, vector of season starting years,
e.g.for season 1995-1996 it would be 1995.}
}
\value{
\code{character()}, vector of seasons suited for the NHL API.
}
\description{
The NHL API wants seasons defined in format
YYYYZZZZ where ZZZZ = YYYY + 1. This is a helper to take
a vector of years in YYYY format and create a vector of
such seasons to be used with the API.
}
\examples{
  nhl_make_seasons(1995:2000)

}
