% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_players_seasons.R
\name{nhl_players_seasons}
\alias{nhl_players_seasons}
\title{Retrieve selected seasons statistics for players}
\usage{
nhl_players_seasons(playerNames, seasons, playerIds = NULL,
  playoffs = FALSE)
}
\arguments{
\item{playerNames}{\code{character()}, player names, not case
sensitive for convenience.}

\item{seasons}{\code{numeric()}, vector of starting years of desired
seasons, in \code{YYYY} format, e.g. \code{1995} for season 1995-1996.
Accepts vectors such as \code{c(1995:2000, 2010)}.
Alternatively, also accepts \code{character()} with seasons in the
format \code{"YYYYZZZZ"}, where ZZZZ = YYYY + 1, e.g. \code{"19951996"}.
This is the format that ultimately gets sent to the API.}

\item{playerIds}{\code{integer()}, ids of the players.}

\item{playoffs}{\code{logical(1)} if \code{FALSE} (default) get the regular
seasons data, if \code{TRUE}, get the data for the playoffs.}
}
\value{
data.frame
}
\description{
Retrieve selected seasons statistics for players
}
\examples{
\dontrun{
 nhl_players_seasons(
   playerIds = c(8451101, 8458554),
   seasons = "19951996",
   playoffs = TRUE
 )
}
}
