% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_teams.R
\name{nhl_teams_rosters}
\alias{nhl_teams_rosters}
\title{Get rosters for teams}
\usage{
nhl_teams_rosters(teamIds = NULL, seasons = NULL)
}
\arguments{
\item{teamIds}{\code{integer()}, ids of the teams or \code{NULL} (default)
for all teams. As of end of 2019, the valid team ids seem to be
in the \code{1:54} range.}

\item{seasons}{\code{numeric()}, vector of starting years of desired
seasons in \code{YYYY} format, e.g. \code{1995} for season 1995-1996.
Accepts vectors such as \code{c(1995:2000, 2010)}.

Alternatively, also accepts \code{character()} with seasons in the
format \code{"YYYYZZZZ"}, where ZZZZ = YYYY + 1, e.g. \code{"19951996"}.
This is the format that ultimately gets sent to the NHL API.

Some API endpoints, notably \code{seasons} exposed via \code{nhl_seasons()}
also allow the value \code{"current"} to passed.}
}
\value{
data.frame, with an element called \code{roster.roster}
that in itself is a data.frame with the roster data.
}
\description{
Get rosters for teams
}
\examples{
 \dontrun{
  # Current rosters for all teams
  nhl_teams_rosters()

  # Rosters for all teams for past seasons
  nhl_teams_rosters(seasons = c("19931994", "19931994"))

  # Roster for Devils and Islanders
  nhl_teams_rosters(
    teamIds = 1:2,
    seasons = c("19931994", "19931994")
  )
}
}
