% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_teams.R
\name{nhl_teams_stats}
\alias{nhl_teams_stats}
\title{Get team statistics per seasons}
\usage{
nhl_teams_stats(teamIds = NULL, seasons = NULL)
}
\arguments{
\item{teamIds}{\code{integer()}, ids of the teams or \code{NULL} (default)
for all teams. As of end of 2019, the valid team ids seem to be
in the \code{1:54} range.}

\item{seasons}{\code{character()}, seasons in the format \code{"YYYYZZZZ"},
where ZZZZ = YYYY + 1, e.g. \code{"19951996"}.}
}
\value{
data.frame with seasons statistics for the selected team(s),
one row per each team and season combination
}
\description{
Get team statistics per seasons
}
\examples{
\dontrun{
  # All teams, current seasons
  nhl_teams_stats()

  # 2 teams, 3 seasons
  nhl_teams_stats(1:2, c("20052006", "20062007", "20072008"))
}
}
