% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_url.R
\name{nhl_url}
\alias{nhl_url}
\title{Create an NHL API URL}
\usage{
nhl_url(endPoint = NULL, suffixes = NULL, params = NULL,
  baseUrl = getOption("nhlapi_baseurl"))
}
\arguments{
\item{endPoint}{\code{character(1)} the endpoint.}

\item{suffixes}{\code{list} of suffixes that will be
concatenated to the end of the urls, separated by
\code{/}.}

\item{params}{\code{named list} of parameters that will be concatenated
to the end of the url after \code{?}.}

\item{baseUrl}{\code{character(1)}, base location of the NHL API.}
}
\value{
\code{character()}, the created URLs.
}
\description{
Create an NHL API URL
}
\examples{
  nhl_url("people", "8477474")

}
