% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_url.R
\name{nhl_url_add_params}
\alias{nhl_url_add_params}
\title{Add parameters to URLs}
\usage{
nhl_url_add_params(url, params = NULL)
}
\arguments{
\item{url}{\code{character()}, vector of URLs.}

\item{params}{\verb{named list()} of parameters that will be
concatenated to the end of the URLs after \verb{?}. Parameters
can have multiple values, in which case multiple URLs
are created. Multiple parameters are separated by \code{&}.}
}
\value{
\code{character()}, URLs with parameters added. Same
length as all the combinations of \code{url} and \code{params}.
}
\description{
Add parameters to URLs
}
