% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_url.R
\name{nhl_url_add_suffixes}
\alias{nhl_url_add_suffixes}
\title{Add suffixes to urls}
\usage{
nhl_url_add_suffixes(url, suffixes)
}
\arguments{
\item{url}{\code{character()} vector of urls.}

\item{suffixes}{\code{list()} of suffixes that will be
concatenated to the end of the urls, separated by
\code{/}.}
}
\value{
\code{character()} of the same length as all the
combinations of \code{url} and \code{suffixes}.
}
\description{
Add suffixes to urls
}
