% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_players.R
\name{nhl_url_players}
\alias{nhl_url_players}
\title{Create an NHL API URL for players}
\usage{
nhl_url_players(playerIds)
}
\arguments{
\item{playerIds}{\code{integer()}, ids of the players.}
}
\value{
\code{character()} of same length as \code{playerIds}
}
\description{
Create an NHL API URL for players
}
\examples{
\dontrun{
  nhl_url_players(c(8477474, 8477475))
}

}
