% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_players_seasons.R
\name{nhl_url_players_allseasons}
\alias{nhl_url_players_allseasons}
\title{Create an NHL API URL for all players' seasons statistics}
\usage{
nhl_url_players_allseasons(playerIds)
}
\arguments{
\item{playerIds}{\code{integer()}, vector of one or more ids of the
players. The ids correspond to the ids expected by the NHL
API's people endpoint. For most cases the \code{playerNames} argument
can be provided for more convenient usage.}
}
\description{
Create an NHL API URL for all players' seasons statistics
}
\examples{
\dontrun{
  # Joe Sakic, all seasons
  nhl_url_players_allseasons(8451101L)
}
}
