% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_players_seasons.R
\name{nhl_url_players_seasons}
\alias{nhl_url_players_seasons}
\title{Create an NHL API URL for  players' seasons statistics}
\usage{
nhl_url_players_seasons(playerIds, seasons, playoffs = FALSE)
}
\arguments{
\item{playerIds}{\code{integer()}, vector of one or more ids of the
players. The ids correspond to the ids expected by the NHL
API people endpoint. For most cases the \code{playerNames} argument
can be provided for more convenient usage.}

\item{seasons}{\code{numeric()}, \code{integer()} or \code{character()},
vector of starting years of desired seasons in \code{YYYY}
format, e.g. \code{1995} or \code{"1995"} for season 1995-1996.
Accepts vectors such as \code{c(1995:2000, 2010)} to generate
multiple seasons.

Alternatively, also accepts \code{character()} with seasons in the
format \code{"YYYYZZZZ"}, where \code{ZZZZ = YYYY + 1}, e.g. \code{"19951996"}.
This is the format that ultimately gets sent to the NHL API.

Some API endpoints, notably \code{seasons} exposed via \code{nhl_seasons()}
also allow the value \code{"current"} to passed. This value will be
returned unchanged.}

\item{playoffs}{\code{logical(1)} if \code{FALSE} (default) get the regular
seasons data, if \code{TRUE}, get the data for the playoffs.}
}
\description{
Create an NHL API URL for  players' seasons statistics
}
\details{
If multiple players and seasons are provided, urls will
be created for all combinations of players and seasons.
}
\examples{
\dontrun{
  # Joe Sakic, regular season 1995/1996
  nhl_url_players_seasons(8451101L, 1995)
  # Joe Sakic, playoffs 1995/1996, 1996/1997 and 1997/1998
  nhl_url_players_seasons(
    8451101L,
    1995:1997,
    playoffs = TRUE
   )
}
}
