% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_utils.R
\name{util_attributes_to_cols}
\alias{util_attributes_to_cols}
\title{Add attributes as data frame columns}
\usage{
util_attributes_to_cols(lst, df, atrs = c("url", "copyright"))
}
\arguments{
\item{lst}{\code{list}, with attributes to be added as columns to \code{df}.}

\item{df}{\code{data.frame}, onto which new columns containing attributes
of \code{lst} should be added.}

\item{atrs}{\code{character()}, vector of names of attributes
of \code{lst}.}
}
\value{
\code{data.frame}, \code{df} with added columns.
}
\description{
Take attributes with names specified by \code{atrs}
from object \code{lst} and adds their value into columns with the same
name in \code{df}.
}
