% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_utils.R
\name{util_convert_minsonice}
\alias{util_convert_minsonice}
\title{Convert "mins:secs" character to numeric minutes}
\usage{
util_convert_minsonice(chr, splitter = ":")
}
\arguments{
\item{chr}{\code{character()} vector in format \code{"mins:secs"}.}

\item{splitter}{\code{character(1)}, string that splits
minutes and seconds in elements of \code{chr}.}
}
\value{
\code{numeric()} vector of times in minutes.
}
\description{
Convert "mins:secs" character to numeric minutes
}
\examples{
\dontrun{
   nhlstats:::MakeMinsOnIce(c("20:00", "1500:30"))
}
}
