% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_utils.R
\name{util_md5sum_str}
\alias{util_md5sum_str}
\title{Get MD5 hash for a character vector}
\usage{
util_md5sum_str(x)
}
\arguments{
\item{x}{\code{character()}, vector to compute the MD5 for.}
}
\value{
\code{character(1)}, MD5 hash of a text file
created from \code{x} using \code{\link[=writeChar]{writeChar()}}.
}
\description{
Writes \code{x} to a temporary file
using \code{writeChar()} and computes the \code{md5sum()}
on that file, removing the file afterwards.
}
\examples{
\dontrun{
  util_md5sum_str("test")
}
}
