% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_utils.R
\name{util_process_minsonice}
\alias{util_process_minsonice}
\title{Convert time columns from "mins:secs" to numeric minutes}
\usage{
util_process_minsonice(df, patt = "timeOn|TimeOn")
}
\arguments{
\item{df}{\code{data.frame}.}

\item{patt}{\code{character(1)}, pattern to match column names that
contain time information in "mm:ss" format.}
}
\value{
\code{data.frame}, with time columns converted from
"mm:ss" characters to numeric minutes.
}
\description{
Convert time columns from "mins:secs" to numeric minutes
}
